+++ packages/thumbprint-react/index.ts
@@ -1,5 +1,5 @@
 export { UserAvatar, EntityAvatar } from './components/Avatar/index';
-export { default as Button, TextButton } from './components/Button/index.jsx';
+export { default as Button, TextButton } from './components/Button/index';
 export { default as ButtonRow } from './components/ButtonRow/index';
 export { BlockList, BlockListItem, BlockListItemLink } from './components/BlockList/index.jsx';
 export { default as Carousel } from './components/Carousel/index';
@@ -12,7 +12,7 @@ export { default as Image } from './components/Image/index';
 export { default as Input, InputIcon, InputClearButton } from './components/Input/index';
 export { default as InputRow, InputRowContext } from './components/InputRow/index.jsx';
 export { default as Label } from './components/Label/index';
-export { default as Link, ThemedLink } from './components/Link/index.jsx';
+export { default as Link, ThemedLink } from './components/Link/index';
 export { List, ListItem } from './components/List/index';
 export { default as LoaderDots } from './components/LoaderDots/index';
 export { default as ModalBase } from './components/ModalBase/index.jsx';