+++ packages/thumbprint-react/index.ts
@@ -13,7 +13,7 @@ export { default as Image } from './components/Image/index';
 export { default as Input, InputIcon, InputClearButton } from './components/Input/index';
 export { default as InputRow, InputRowContext } from './components/InputRow/index.jsx';
 export { default as Label } from './components/Label/index';
-export { default as Link, ThemedLink } from './components/Link/index.jsx';
+export { default as Link, ThemedLink } from './components/Link/index';
 export { List, ListItem } from './components/List/index';
 export { default as LoaderDots } from './components/LoaderDots/index';
 export { default as ModalBase } from './components/ModalBase/index.jsx';