+++ packages/thumbprint-react/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Added
+
+-   [Minor] Create `Alert` component in React (TypeScript).
+
 ## 10.0.1 - 2019-10-21
 
 ### Fixed