+++ packages/thumbprint-react/CHANGELOG.md
@@ -6,6 +6,7 @@
 
 -   [Patch] Rewrite `Button` component in TypeScript.
 -   [Patch] Rewrite `UIAction` component in TypeScript.
+-   [Patch] Rewrite `Link` component in TypeScript.
 
 ### Fixed