+++ packages/thumbprint-react/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Changed
+
+-   [Patch] Rewrite internal utility function and components in TypeScript.
+
 ### Fixed
 
 -   [Patch] Update Type syntax for `Image` and `ServiceCardImage` to fix missing API docs.